moved types around to make it easier to control in one place and propagate downstream
